[[ TileCache ]] Add GLHardwareBuffer API and Android implementation
This patch adds an API for obtaining a hardware buffer that can be bound
to a GL texture and locked for direct access to the memory buffer.

This is implemented on Android using the AHardwareBuffer C api. The API
functions are dynamically loaded to avoid linkage failures when
attempting to run on Android versions that don't support the APIs.