[[ Bug 22586 ]] Fix buffer overrun in native split
This patch fixes a buffer overrun in the native split codepath where
adding the key length to the ptr used to extract the element could result
in a start pointer higher than the end ptr. Subtracting from the end ptr to
find the length resulted in a very high length due to unsignedness thus causing
invalid access.