[[ Bug 22960 ]] Use simctl to deploy to iOS simulator
This patch uses the simctl tool to deploy to iOS simulator. This
approach is more stable than the mothods used in the reviphoneproxy external,
and also fixes a number of issues such as deployment when the simulator is not up and running,
or deployment when the simulator app is open but no device is booted.

Tested with iOS 13.3 and iOS 14.1 simulator, in all possible combinations of the simulator status.