[[ Bug 22787 ]] Fix crash when reopening modal stack in "resumeStack" handler
This patch fixes a crash when opening, closing, then reopening a modal
stack in the "resumeStack" handler of another stack.

Previously, the "resumeStack" message would be dispatched while in the
middle of the NSWindow::orderOut message. Starting a new modal session
for the NSWindow in that handler was raising an Obj-C exception leading
to crashing & instability. We can avoid that by deferring the focus and
unfocus messages until after the NSWindow::orderOut message has
completed.