[[ Bug 22896 ]] Fix click location for sheet dialogs on MacOS Big Sur
This patch fixes an issue on MacOS Big Sur where setting the rect of an
answer dialog opened as sheet would fail, leaving the content rect in an
inconsistent state, leading to an incorrect offset added to the mouse
location.
This was preventing the dialog buttons from responding to mouse clicks.