[[ Bug 23053 ]] Fix frozen color picker launched from modal stack on macOS
This patch fixes a bug on macOS where a color picker launched from a
modal dialog would be unresponsive to input events, leaving the engine
stuck in an infinite loop. This is fixed by dispatching events from
the event queue in the usual way rather than by running the modal
session.