[[ Bug 23106 ]] Do not remove hidden native layers on Android
This patch changes the behavior of hiding widgets with a native layer on
Android so that the native view has its visibility set to invisible rather
then it being removed from the view hierarchy entirely. The result of this
is that native views which require window context to function (e.g. VideoView)
no longer break when made invisible and then visible again.