[[ Bug 21906 ]] Fix copying data from clipboard via IStream objects on Windows
This patch fixes bug 21906, where pasting image data from some
applications would fail, resulting in an empty image control.
This happens when the data is transferred via an IStream object, and is
fixed by ensuring the stream pointer is reset to the beginning of the
stream before reading from it.