[[ Bug 23148 ]] Cleanup access to keyword tables
This patch cleans up access to the keyword tables, used by the script parser
and syntax such as `the commandNames`. The tables are now accessible via
global variables `MCkeywordtablepointers` and `MCkeywordtablesizes` which are
defined in `scriptpt.h`; and all direct access to the underlying tables in
other files have been updated to use them.