[[ Bug 23038 ]] Fix creation of color profiles on macOS
This patch fixes an issue with the creation of color profiles from
embedded ICC information. Previously the NoCopy CFDataRef constructor
was being used to wrap the raw bytes to be passed to the relevant
CGColorSpaceRef constructor. However, the data wrapped in this way
could be released at any time, and in versions of macOS since
Mojave it appears that the OS defers copying the data thus causing
strange effects when the color profile is eventually applied. The
problem has been fixed by using the normal CFDataRef constructor
which copies the provided bytes.